src/backend/requirements.txt
bcrypt==4.2.1
fastapi==0.103.0
motor==3.6.1
numpy==2.2.1
pandas==2.2.3
pydantic==2.10.5
pymongo==4.9.2
python-dotenv==1.0.0
uvicorn==0.23.0